Release/v0.1.0

+++ contracts/libraries/CumulativeNormalDistribution.sol
@@ -15,8 +15,8 @@ library CumulativeNormalDistribution {
         // where p = 0.3275911,
         // a1 = 0.254829592, a2 = −0.284496736, a3 = 1.421413741, a4 = −1.453152027, a5 = 1.061405429
         int128 p = 0x53dd02a4f5ee2e46;
-        int128 one = uint(1).fromUInt();
-        int128 two = uint(2).fromUInt();
+        int128 one = uint256(1).fromUInt();
+        int128 two = uint256(2).fromUInt();
         int128 a1 = 0x413c831bb169f874;
         int128 a2 = -0x48d4c730f051a5fe;
         int128 a3 = 0x16a09e667f3bcc908;
@@ -43,13 +43,7 @@ library CumulativeNormalDistribution {
         int128 b0 = 0x21D0A04B0E9BA0F0; // 0.132089632
         int128 b1 = -0xC2BF5D74C7247680; // -0.760732499
         // fcentral(p) = q * (a2 + (a1r + a0) / (r^2 + b1r +b0))
-        int128 result = q.mul(
-            a2.add(
-                (a1.mul(r).add(a0)).div(
-                    (r.pow(2).add(b1.mul(r)).add(b0))
-                )
-            )
-        );
+        int128 result = q.mul(a2.add((a1.mul(r).add(a0)).div((r.pow(2).add(b1.mul(r)).add(b0)))));
         return result;
     }
